Fix broken test
Not sure how this ever worked. The code being tested threw a null error for 'roles' when checking `member.djs.roles.has`, which was changed in the previous commit. But what makes me confused is that all the previous commit does is change this line from being in a separate function to being written here directly. The same check would have been done before, and 'roles' would have been just as null. I don't really want to go back and figure it out so here we are, it works now.